# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook AutoPoster API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List connected Facebook pages
 */
export const ListPagesResponseItem = zod.object({
  id: zod.number(),
  pageId: zod.string(),
  name: zod.string(),
  category: zod.string().nullish(),
  avatarUrl: zod.string().nullish(),
  accessToken: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListPagesResponse = zod.array(ListPagesResponseItem);

/**
 * @summary Connect a Facebook page using access token
 */
export const ConnectPageBody = zod.object({
  accessToken: zod
    .string()
    .describe(
      "Facebook user access token — we will fetch page details automatically",
    ),
});

/**
 * @summary Disconnect a Facebook page
 */
export const DisconnectPageParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all posts
 */
export const ListPostsQueryParams = zod.object({
  status: zod.enum(["draft", "scheduled", "published", "failed"]).optional(),
  pageId: zod.coerce.number().optional(),
});

export const ListPostsResponseItem = zod.object({
  id: zod.number(),
  pageId: zod.number(),
  pageName: zod.string().nullish(),
  title: zod.string().nullish(),
  postType: zod.enum(["text", "image", "video"]),
  caption: zod.string(),
  imageUrl: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  status: zod.enum(["draft", "scheduled", "published", "failed"]),
  scheduledAt: zod.coerce.date().nullish(),
  publishedAt: zod.coerce.date().nullish(),
  facebookPostId: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListPostsResponse = zod.array(ListPostsResponseItem);

/**
 * @summary Create a new post
 */
export const CreatePostBody = zod.object({
  pageId: zod.number(),
  title: zod.string().nullish(),
  postType: zod.enum(["text", "image", "video"]).optional(),
  caption: zod.string(),
  imageUrl: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  scheduledAt: zod.coerce.date().nullish(),
});

/**
 * @summary Get a post by ID
 */
export const GetPostParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPostResponse = zod.object({
  id: zod.number(),
  pageId: zod.number(),
  pageName: zod.string().nullish(),
  title: zod.string().nullish(),
  postType: zod.enum(["text", "image", "video"]),
  caption: zod.string(),
  imageUrl: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  status: zod.enum(["draft", "scheduled", "published", "failed"]),
  scheduledAt: zod.coerce.date().nullish(),
  publishedAt: zod.coerce.date().nullish(),
  facebookPostId: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a post
 */
export const UpdatePostParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePostBody = zod.object({
  title: zod.string().nullish(),
  postType: zod.enum(["text", "image", "video"]).optional(),
  caption: zod.string().optional(),
  imageUrl: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  scheduledAt: zod.coerce.date().nullish(),
  status: zod.enum(["draft", "scheduled"]).optional(),
});

export const UpdatePostResponse = zod.object({
  id: zod.number(),
  pageId: zod.number(),
  pageName: zod.string().nullish(),
  title: zod.string().nullish(),
  postType: zod.enum(["text", "image", "video"]),
  caption: zod.string(),
  imageUrl: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  status: zod.enum(["draft", "scheduled", "published", "failed"]),
  scheduledAt: zod.coerce.date().nullish(),
  publishedAt: zod.coerce.date().nullish(),
  facebookPostId: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a post
 */
export const DeletePostParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Publish a post immediately to Facebook
 */
export const PublishPostParams = zod.object({
  id: zod.coerce.number(),
});

export const PublishPostResponse = zod.object({
  id: zod.number(),
  pageId: zod.number(),
  pageName: zod.string().nullish(),
  title: zod.string().nullish(),
  postType: zod.enum(["text", "image", "video"]),
  caption: zod.string(),
  imageUrl: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  status: zod.enum(["draft", "scheduled", "published", "failed"]),
  scheduledAt: zod.coerce.date().nullish(),
  publishedAt: zod.coerce.date().nullish(),
  facebookPostId: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Generate an AI image using Hugging Face free inference
 */
export const GenerateImageBody = zod.object({
  prompt: zod.string().describe("Text prompt to generate an image from"),
  style: zod
    .string()
    .nullish()
    .describe("Style hint e.g. photorealistic, cartoon, watercolor"),
});

export const GenerateImageResponse = zod.object({
  imageUrl: zod.string(),
  prompt: zod.string(),
});

/**
 * @summary Get dashboard summary statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalPages: zod.number(),
  totalPosts: zod.number(),
  scheduledPosts: zod.number(),
  publishedPosts: zod.number(),
  draftPosts: zod.number(),
  failedPosts: zod.number(),
});

/**
 * @summary Get the 5 most recent posts across all pages
 */
export const GetRecentPostsResponseItem = zod.object({
  id: zod.number(),
  pageId: zod.number(),
  pageName: zod.string().nullish(),
  title: zod.string().nullish(),
  postType: zod.enum(["text", "image", "video"]),
  caption: zod.string(),
  imageUrl: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  status: zod.enum(["draft", "scheduled", "published", "failed"]),
  scheduledAt: zod.coerce.date().nullish(),
  publishedAt: zod.coerce.date().nullish(),
  facebookPostId: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetRecentPostsResponse = zod.array(GetRecentPostsResponseItem);

/**
 * @summary Get upcoming scheduled posts
 */
export const GetUpcomingPostsResponseItem = zod.object({
  id: zod.number(),
  pageId: zod.number(),
  pageName: zod.string().nullish(),
  title: zod.string().nullish(),
  postType: zod.enum(["text", "image", "video"]),
  caption: zod.string(),
  imageUrl: zod.string().nullish(),
  audioUrl: zod.string().nullish(),
  status: zod.enum(["draft", "scheduled", "published", "failed"]),
  scheduledAt: zod.coerce.date().nullish(),
  publishedAt: zod.coerce.date().nullish(),
  facebookPostId: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetUpcomingPostsResponse = zod.array(GetUpcomingPostsResponseItem);